split the profile links into a separate comp
